pyproject.toml
[project]
name = "pdfconvert"
version = "0.2.0"
description = "PDF to Markdown converter using LM Studio vision models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.40.0",
    "requests>=2.32.0",
    "pdf2image>=1.17.0",
    "pillow>=12.1.1",
]

[project.scripts]
pdfconvert = "pdfconvert.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pdfconvert"]
